refactor: Задача спринта №17 исправления по замечаниям ревьювера - интерфейсы репозиториев перенесены в слой domain
