+++ app/src/main/java/com/practicum/playlistmaker/main/di/MainDataModule.kt
@@ -1,6 +1,6 @@
 package com.practicum.playlistmaker.main.di
 
-import com.practicum.playlistmaker.main.data.MainRepository
+import com.practicum.playlistmaker.main.domain.MainRepository
 import com.practicum.playlistmaker.main.data.impl.MainRepositoryImpl
 import org.koin.android.ext.koin.androidApplication
 import org.koin.dsl.module
