+++ app/src/main/java/com/practicum/playlistmaker/main/domain/MainRepository.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.main.data
+package com.practicum.playlistmaker.main.domain
 
 interface MainRepository {
     fun openSearchActivity()
