+++ app/src/main/java/com/practicum/playlistmaker/main/domain/impl/MainInteractorImpl.kt
@@ -1,6 +1,6 @@
 package com.practicum.playlistmaker.main.domain.impl
 
-import com.practicum.playlistmaker.main.data.MainRepository
+import com.practicum.playlistmaker.main.domain.MainRepository
 import com.practicum.playlistmaker.main.domain.MainInteractor
 
 class MainInteractorImpl(private val mainRepository: MainRepository) : MainInteractor {
