+++ app/src/main/java/com/practicum/playlistmaker/player/domain/PlayerRepository.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.player.data
+package com.practicum.playlistmaker.player.domain
 
 import com.practicum.playlistmaker.player.domain.models.PlayerState
 import com.practicum.playlistmaker.search.domain.models.Track
