+++ app/src/main/java/com/practicum/playlistmaker/search/data/impl/SearchHistoryRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.practicum.playlistmaker.search.data.impl
 
-import com.practicum.playlistmaker.search.data.SearchHistoryRepository
+import com.practicum.playlistmaker.search.domain.SearchHistoryRepository
 import com.practicum.playlistmaker.search.domain.models.Track
 import com.practicum.playlistmaker.search.domain.models.TracksList
 import com.practicum.playlistmaker.utils.data.SharedPreferencesRepository
