+++ app/src/main/java/com/practicum/playlistmaker/search/domain/TrackRepository.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.search.data
+package com.practicum.playlistmaker.search.domain
 
 import com.practicum.playlistmaker.search.domain.models.Track
