+++ app/src/main/java/com/practicum/playlistmaker/settings/di/SettingsDataModule.kt
@@ -1,6 +1,6 @@
 package com.practicum.playlistmaker.settings.di
 
-import com.practicum.playlistmaker.settings.data.SettingsRepository
+import com.practicum.playlistmaker.settings.domain.SettingsRepository
 import com.practicum.playlistmaker.settings.data.impl.SettingsRepositoryImpl
 import com.practicum.playlistmaker.settings.data.models.ApplicationSettings
 import com.practicum.playlistmaker.utils.data.SharedPreferencesRepository
