+++ app/src/main/java/com/practicum/playlistmaker/settings/domain/SettingsRepository.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.settings.data
+package com.practicum.playlistmaker.settings.domain
 
 import com.practicum.playlistmaker.settings.data.models.ApplicationSettings
