+++ app/src/main/java/com/practicum/playlistmaker/settings/domain/impl/SettingsInteractorImpl.kt
@@ -1,6 +1,6 @@
 package com.practicum.playlistmaker.settings.domain.impl
 
-import com.practicum.playlistmaker.settings.data.SettingsRepository
+import com.practicum.playlistmaker.settings.domain.SettingsRepository
 import com.practicum.playlistmaker.settings.data.models.ApplicationSettings
 import com.practicum.playlistmaker.settings.domain.SettingsInteractor
