+++ app/src/main/java/com/practicum/playlistmaker/sharing/data/impl/SharingRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.practicum.playlistmaker.sharing.data.impl
 
 import android.content.Context
 import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.sharing.data.SharingRepository
+import com.practicum.playlistmaker.sharing.domain.SharingRepository
 import com.practicum.playlistmaker.sharing.domain.models.EmailData
 
 class SharingRepositoryImpl(private val context: Context) : SharingRepository {
