+++ app/src/main/java/com/practicum/playlistmaker/sharing/di/SharingDataModule.kt
@@ -1,7 +1,7 @@
 package com.practicum.playlistmaker.sharing.di
 
 import com.practicum.playlistmaker.sharing.data.ExternalNavigator
-import com.practicum.playlistmaker.sharing.data.SharingRepository
+import com.practicum.playlistmaker.sharing.domain.SharingRepository
 import com.practicum.playlistmaker.sharing.data.impl.ExternalNavigatorImpl
 import com.practicum.playlistmaker.sharing.data.impl.SharingRepositoryImpl
 import org.koin.android.ext.koin.androidContext
